+++ spb/backends/matplotlib/matplotlib.py
@@ -412,6 +412,9 @@ class MatplotlibBackend(Plot):
                     self.plt.FuncFormatter(
                         self.y_ticks_formatter.MB_func_formatter()))
 
+        for f in self.hooks:
+            f(self._fig, self._ax)
+
     def _set_axes_texts(self):
         title, xlabel, ylabel, zlabel = self._get_title_and_labels()